Move MC_ALTT to userspace for cross-board support
