+++ users/bcat/bcat.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "quantum.h"
+
+enum user_keycodes {
+    MC_ALTT = SAFE_RANGE,
+    KEYMAP_SAFE_RANGE,
+};
